requirements.txt
# Chronos Parser

pyodbc==4.0.39
python-dotenv==1.0.0 